Excepciones terminadas e inicio de arhcivos en Decorador

+++ PROYECTO 1/DecoradorAbarrote.h
@@ -13,9 +13,9 @@ private:
 	double peso;
 	double precioCosto;
 	string categoria;
-	double cantidad;
+	int cantidad;
 public:
-	DecoradorAbarrote(ComponenteAbstracto*, Perecedero*, string, string, string, string, bool, double, double,double);
+	DecoradorAbarrote(ComponenteAbstracto*, Perecedero*, string, string, string, string, bool, double, double,int);
 	DecoradorAbarrote(const DecoradorAbarrote&);
 	virtual ~DecoradorAbarrote();
 	virtual ComponenteAbstracto* getSiguiente();
@@ -26,6 +26,9 @@ public:
 	virtual double getCantidad();
 	virtual ComponenteAbstracto* clonar() const;
 
+	//Archivos
+	//virtual void guardarComponenteAbstracto(ofstream&);
+	//static DecoradorAbarrote* leerDecoAbarrote(ifstream&);
 
 	//Sobrecargas
 	friend ostream& operator<<(ostream&, DecoradorAbarrote&);
