+++ PROYECTO 1/DecoradorEmbutido.h
@@ -15,10 +15,10 @@ private:
     double precioCosto;
     bool nacional;
     string categoria;
-    double cantidad;
+    int cantidad;
 public:
 
-    DecoradorEmbutido(ComponenteAbstracto*, Empaque*, Perecedero*, string, string, string, string, string, string, double, double, bool, double);
+    DecoradorEmbutido(ComponenteAbstracto*, Empaque*, Perecedero*, string, string, string, string, string, string, double, double, bool, int);
 	DecoradorEmbutido(const DecoradorEmbutido&);
 	virtual ~DecoradorEmbutido();
 	virtual ComponenteAbstracto* getSiguiente();
@@ -29,6 +29,9 @@ public:
     virtual double getCantidad();
     virtual ComponenteAbstracto* clonar() const;
 
+    //Archivos
+    //virtual void guardarComponenteAbstracto(ofstream&);
+    //static DecoradorEmbutido* leerDecoEmbutido(ifstream&);
 
     //Sobrecargas
     friend ostream& operator<<(ostream& o, DecoradorEmbutido&);
